+++ CHANGELOG.md
@@ -49,6 +49,8 @@ file sha1`, `info file all`. Now all information is listed under `info file`.
 - `testing` setting. It was just a hack to be able to test `byebug`. Nobody was
 supposed to actually use it!
 - `var class` command, just use Ruby (`self.class.class_variables`).
+- `p` command, just use `eval`, or just type your expression and `byebug` will
+autoevaluate it.
 
 ## 3.5.1 - 2014-09-29
 ### Fixed